Bugfix: prevent empty or whitespace questions/answers
+ 5 additional tests

+++ cogs/qanda.py
@@ -31,6 +31,16 @@ class Qanda(commands.Cog):
             await ctx.message.delete()
             return
 
+        if not qs or qs.isspace():
+            await ctx.author.send("Please enter a valid question.")
+            await ctx.message.delete()
+            return
+
+        if len(qs) <= 2:
+            await ctx.author.send('Question too short.')
+            await ctx.message.delete()
+            return
+
         # get author
         if anonymous == '':
             author = ctx.message.author.id
@@ -94,7 +104,7 @@ class Qanda(commands.Cog):
     @commands.command(name='answer',
                       help='Answer question. Please put answer text in quotes. Add *anonymous* or *anon* if desired.'
                            'EX: $answer 1 /"Oct 12/" anonymous')
-    async def answer(self, ctx, num, ans, anonymous=''):
+    async def answer(self, ctx, num, ans: str, anonymous=''):
         ''' answer the specific question '''
         # make sure to check that this is actually being asked in the Q&A channel
         if not ctx.channel.name == 'q-and-a':
@@ -102,16 +112,15 @@ class Qanda(commands.Cog):
             await ctx.message.delete()
             return
 
-        # to stop SQL from freezing. Only allows valid numbers.
-        if not re.match(r'^([1-9]\d*|0)$', num):
-            await ctx.author.send('Please include a valid question number. EX: $answer 1 /"Oct 12/" anonymous')
+        if not ans or ans.isspace():
+            await ctx.author.send("Please enter a valid answer.")
             await ctx.message.delete()
             return
 
-        #if not isinstance(ans, str):
-            #await ctx.author.send('Please include a valid answer. EX: $answer 1 /"Oct 12/" anonymous')
-           # await ctx.message.delete()
-           # return
+#        if len(ans) == 0:
+#            await ctx.author.send('Answer too short.')
+#            await ctx.message.delete()
+#            return
 
         # get author
         if anonymous == '':
@@ -122,6 +131,12 @@ class Qanda(commands.Cog):
             author = None
         else:
             await ctx.author.send('Unknown input for *anonymous* option. Please type **anonymous**, **anon**, or leave blank.')
+            await ctx.message.delete()
+            return
+
+        # to stop SQL from freezing. Only allows valid numbers.
+        if not re.match(r'^([1-9]\d*|0)$', num):
+            await ctx.author.send('Please include a valid question number. EX: $answer 1 /"Oct 12/" anonymous')
             await ctx.message.delete()
             return
